+++ wshop/src/main/java/com/wshop/entity/MatchColorWork.java
@@ -15,7 +15,7 @@ public class MatchColorWork {
 
     private String customerCode;
 
-    private Double number;
+    private Integer number;
 
     private String machineCode;
 
@@ -77,11 +77,11 @@ public class MatchColorWork {
         this.customerCode = customerCode == null ? null : customerCode.trim();
     }
 
-    public Double getNumber() {
+    public Integer getNumber() {
         return number;
     }
 
-    public void setNumber(Double number) {
+    public void setNumber(Integer number) {
         this.number = number;
     }
