+++ wshop/src/main/java/com/wshop/entity/Order.java
@@ -15,7 +15,7 @@ public class Order {
 
     private String material;
 
-    private Double number;
+    private Integer number;
 
     private String customer;
 
@@ -73,11 +73,11 @@ public class Order {
         this.material = material == null ? null : material.trim();
     }
 
-    public Double getNumber() {
+    public Integer getNumber() {
         return number;
     }
 
-    public void setNumber(Double number) {
+    public void setNumber(Integer number) {
         this.number = number;
     }
